Move Header to GhBase for 625XT/505/580

+++ Device/Gh505Device.cs
@@ -64,7 +64,7 @@ namespace ZoneFiveSoftware.SportTracks.Device.Globalsat
             foreach (Gh505Packet.TrackFileHeader header in tracks)
             {
                 totalPoints += header.TrackPointCount;
-                trackIndexes.Add(header.TrackPointIndex);
+                trackIndexes.Add((Int16)header.TrackPointIndex);
             }
             float pointsRead = 0;
